+++ src/code/PSRepositoryInfo.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Management.Automation;
+using System.Net;
 
 namespace Microsoft.PowerShell.PSResourceGet.UtilClasses
 {
@@ -100,9 +101,36 @@ namespace Microsoft.PowerShell.PSResourceGet.UtilClasses
 
         internal bool IsMARRepository()
         {
-            return (ApiVersion == APIVersion.ContainerRegistry && Uri.Host.StartsWith("mcr.microsoft") );
+            return (ApiVersion == APIVersion.ContainerRegistry && Uri.Host.StartsWith("mcr.microsoft"));
         }
 
+        public static bool IsContainerRegistry(string uri)
+        {
+            if (uri.EndsWith(".azurecr.io") || uri.EndsWith(".azurecr.io/") || uri.Contains("mcr.microsoft.com") || uri.StartsWith("mcr.microsoft"))
+            {
+                return true;
+            }
+
+            return false;
+        }
+        
+        public static NetworkCredential SetNetworkCredentials(PSRepositoryInfo repository, NetworkCredential networkCredential, PSCmdlet cmdletPassedIn)
+        {
+            NetworkCredential networkCreds = new NetworkCredential();
+            if (repository.CredentialProvider.Equals(PSRepositoryInfo.CredentialProviderType.AzArtifacts))
+            {
+                cmdletPassedIn.WriteVerbose("Setting credential provider network credentials");
+                networkCreds = Utils.SetCredentialProviderNetworkCredential(repository, networkCredential, cmdletPassedIn);
+            }
+            else
+            {
+                cmdletPassedIn.WriteVerbose("Setting Secret Management network credentials");
+                networkCreds = Utils.SetSecretManagementNetworkCredential(repository, networkCredential, cmdletPassedIn);
+            }
+
+            return networkCreds;
+        }     
+        
         #endregion
     }
 }
